Better detection for IE11
IE11 on Win 8.1 and newer does not always seem to populate itself as MSIE 11, so the detection is actually detecting the Mozilla string. But it sends out rv:11.0 so we will scan for this instead.

The Trident/x.0 detection will not be re-added, since this version does not tell us with absolute certainty, which Internet Explorer is really running.

IE10 can have three different Trident versions, see this for reference: http://www.useragentstring.com/pages/Internet%20Explorer/