Fix counter splitting
The splitting of the counter value[1] creates an array with an excess
element, namely an empty string.  This causes a "A non well formed
numeric value encountered " notice as of PHP 7.1 in the following
`foreach` loop.  This can be easily avoided by using `str_split()`,
which also simplifies the code.

We also simplify the zero padding immediately before the splitting code
by using `str_pad()`.

[1] <https://github.com/schmatzler/crazystat/blob/09ac03f32e7c918457ee7432819b44e05719f29f/zaehler/src/create_counter_image.php#L81-L82>